Move the cache! method over to jbuilder_template where we can catch the partial! nested as well.
